Move App() component to separate file
